More code changes to x32 variables

+++ include/dynamic_placeholder_text_util.h
@@ -4,9 +4,9 @@
 #include "global.h"
 
 void DynamicPlaceholderTextUtil_Reset(void);
-void DynamicPlaceholderTextUtil_SetPlaceholderPtr(u8 idx, const u8 *ptr);
+void DynamicPlaceholderTextUtil_SetPlaceholderPtr(u32 idx, const u8 *ptr);
 u8 *DynamicPlaceholderTextUtil_ExpandPlaceholders(u8 *dest, const u8 *src);
-const u8 *DynamicPlaceholderTextUtil_GetPlaceholderPtr(u8 idx);
-u8 GetColorFromTextColorTable(u16 graphicId);
+const u8 *DynamicPlaceholderTextUtil_GetPlaceholderPtr(u32 idx);
+u32 GetColorFromTextColorTable(u32 graphicId);
 
 #endif //GUARD_DYNAMIC_PLACEHOLDER_TEXT_UTIL_H
