+++ include/evolution_graphics.h
@@ -5,12 +5,12 @@
 
 void LoadEvoSparkleSpriteAndPal(void);
 
-u8 LaunchTask_PreEvoSparklesSet1(u16 arg0);
-u8 LaunchTask_PreEvoSparklesSet2(void);
-u8 LaunchTask_PostEvoSparklesSet1(void);
-u8 LaunchTask_PostEvoSparklesSet2AndFlash(u16 species);
-u8 LaunchTask_PostEvoSparklesSet2AndFlash_Trade(u16 species);
-u8 LaunchTask_PrePostEvoMonSprites(u8 preEvoSpriteId, u8 postEvoSpriteId);
+u32 LaunchTask_PreEvoSparklesSet1(u32 arg0);
+u32 LaunchTask_PreEvoSparklesSet2(void);
+u32 LaunchTask_PostEvoSparklesSet1(void);
+u32 LaunchTask_PostEvoSparklesSet2AndFlash(u32 species);
+u32 LaunchTask_PostEvoSparklesSet2AndFlash_Trade(u32 species);
+u32 LaunchTask_PrePostEvoMonSprites(u32 preEvoSpriteId, u32 postEvoSpriteId);
 
 #define EvoGraphicsTaskEvoStop  data[8]
