+++ include/vs_seeker.h
@@ -3,7 +3,7 @@
 
 #include "global.h"
 
-void Task_VsSeeker_0(u8 taskId);
+void Task_VsSeeker_0(u32 taskId);
 void ClearRematchStateByTrainerId(void);
 void ClearRematchStateOfLastTalked(void);
 int GetRematchTrainerId(u16 trainerId);
