+++ src/lua_handler.c
@@ -3,7 +3,7 @@
 
 static struct LuaHandler sLuaHandler;
 
-void ProcessLuaAction(u8 action, u32 argument)
+void ProcessLuaAction(u32 action, u32 argument)
 {
 	if (sLuaHandler.status == LUA_HANDLER_STATE_NOTHING)
 	{
@@ -13,9 +13,9 @@ void ProcessLuaAction(u8 action, u32 argument)
 	}
 }
 
-u8 GetLuaActionProcessStatus(void)
+u32 GetLuaActionProcessStatus(void)
 {
-	u8 status = sLuaHandler.status;
+	u32 status = sLuaHandler.status;
 	
 	if (status == LUA_HANDLER_STATE_PROCESS_FINISHED || status == LUA_HANDLER_STATE_PROCESS_FAILED)
 		sLuaHandler.status = LUA_HANDLER_STATE_NOTHING;
